+++ changelog.md
@@ -3,3 +3,10 @@
 - 3.51x Performance Improvement in single_point crossover
 - 2.35x Performance Improvement in two_point crossover
 - Addition of `benchmark/0` function to benchmark your algorithm
+
+# v0.1.5
+- Changed to `libgraph` for Genealogy tree.
+- Added ability to export graph as DOT file.
+- Removed `track_history?` flag.
+- Added 17 Tests
+- Fixed Variance bug